introduce bootevent
Also, make modules export functions.
