+++ lib/devicestatus.js
@@ -42,7 +42,6 @@ function ensureIndexes(name, storage) {
   });
 }
 
-module.exports = {
-  storage: storage,
-  ensureIndexes: ensureIndexes
-};
+
+storage.ensureIndexes = ensureIndexes;
+module.exports = storage;
