+++ lib/entries.js
@@ -198,8 +198,6 @@ function ensureIndexes(name, storage) {
 }
 
 // expose module
-module.exports = {
-  storage: storage,
-  ensureIndexes: ensureIndexes
-};
+storage.ensureIndexes = ensureIndexes;
+module.exports = storage;
