+++ lib/treatments.js
@@ -55,8 +55,6 @@ function ensureIndexes(name, storage) {
     }
   });
 }
+storage.ensureIndexes = ensureIndexes;
+module.exports = storage;
 
-module.exports = {
-  storage: storage,
-  ensureIndexes: ensureIndexes
-};
