+++ package.json
@@ -43,7 +43,8 @@
     "pushover-notifications": "0.2.0",
     "sgvdata": "0.0.2",
     "socket.io": "^0.9.17",
-    "git-rev": "git://github.com/bewest/git-rev.git"
+    "git-rev": "git://github.com/bewest/git-rev.git",
+    "bootevent": "0.0.1"
   },
   "devDependencies": {
     "supertest": "~0.13.0",
