+++ server.js
@@ -26,33 +26,9 @@
 // DB Connection setup and utils
 ///////////////////////////////////////////////////
 
-var software = require('./package.json');
 var env = require('./env')( );
-var pushover = require('./lib/pushover')(env);
+var store = require('./lib/storage')(env);
 
-var entries = require('./lib/entries');
-var treatments = require('./lib/treatments');
-var devicestatus = require('./lib/devicestatus');
-
-var store = require('./lib/storage')(env, function() {
-    console.info("Mongo ready");
-    entries.ensureIndexes(env.mongo_collection, store);
-    treatments.ensureIndexes(env.treatments_collection, store);
-    devicestatus.ensureIndexes(env.devicestatus_collection, store);
-});
-
-
-var express = require('express');
-
-///////////////////////////////////////////////////
-// api and json object variables
-///////////////////////////////////////////////////
-var entriesStorage = entries.storage(env.mongo_collection, store, pushover);
-var settings = require('./lib/settings')(env.settings_collection, store);
-var treatmentsStorage = treatments.storage(env.treatments_collection, store, pushover);
-var devicestatusStorage = devicestatus.storage(env.devicestatus_collection, store);
-var api = require('./lib/api/')(env, entriesStorage, settings, treatmentsStorage, devicestatusStorage);
-var pebble = require('./lib/pebble');
 ///////////////////////////////////////////////////
 
 ///////////////////////////////////////////////////
@@ -60,43 +36,27 @@ var pebble = require('./lib/pebble');
 ///////////////////////////////////////////////////
 var PORT = env.PORT;
 
-var app = express();
-var appInfo = software.name + ' ' + software.version;
-app.set('title', appInfo);
-app.enable('trust proxy'); // Allows req.secure test on heroku https connections.
-
-//if (env.api_secret) {
-//    console.log("API_SECRET", env.api_secret);
-//}
-app.use('/api/v1', api);
-
-// pebble data
-app.get('/pebble', pebble(entriesStorage, devicestatusStorage));
-
-//app.get('/package.json', software);
-
-// define static server
-//TODO: JC - changed cache to 1 hour from 30d ays to bypass cache hell until we have a real solution
-var staticFiles = express.static(env.static_files, {maxAge: 60 * 60 * 1000});
-
-// serve the static content
-app.use(staticFiles);
-
-// Handle errors with express's errorhandler, to display more readable error messages.
-var errorhandler = require('errorhandler');
-//if (process.env.NODE_ENV === 'development') {
-  app.use(errorhandler());
-//}
-
-store(function ready ( ) {
-  var server = app.listen(PORT);
-  console.log('listening', PORT);
-
-  ///////////////////////////////////////////////////
-  // setup socket io for data and message transmission
-  ///////////////////////////////////////////////////
-  var websocket = require('./lib/websocket');
-  var io = websocket(env, server, entriesStorage, treatmentsStorage);
-});
+var bootevent = require('bootevent');
+bootevent( )
+  .acquire(function db (ctx, next) {
+    // initialize db connections
+    store( function ready ( ) {
+      console.log('storage system ready');
+      ctx.store = store;
+      next( );
+    });
+  })
+  .boot(function booted (ctx) {
+    env.store = ctx.store;
+    var app = require('./app')(env);
+    var server = app.listen(PORT);
+    console.log('listening', PORT);
+    ///////////////////////////////////////////////////
+    // setup socket io for data and message transmission
+    ///////////////////////////////////////////////////
+    var websocket = require('./lib/websocket');
+    var io = websocket(env, server, app.entries, app.treatments);
+  })
+;
 
 ///////////////////////////////////////////////////
